Update python 3 version and kafka-python package
With python 3.7 there were changes that meant async became a key word
kafka-python was using async in multiple places which meant they needed
to update their package to work correctly. Our build failed for this
package due to the out of date dependancy.

We have updated the python version for the tests as well as the package
to make sure everything is still working as it should be.

NOTE: Also added docker box and Makefile so it is easier to run tests.