Update makefile so that you can do twine uploads to test and live
Previously the makefile was only uploading to test, which I didn't have
user credentials for, which then wasn't working and I was very baffled.

I have now setup a user account on test pypi so we can upload to test
and live, i've also changed the makefile so it runs a twine package
check after it builds the package for quick feedback as to weather it
will upload and detailed information if anything goes wrong.